action: use int64 for action IDs to fix overflow on 32-bit platforms
Action IDs have exceeded math.MaxInt32 (2,147,483,647). On 32-bit
platforms (e.g. linux/arm), Go's int type is 32 bits, causing
json.Unmarshal to fail when decoding API responses containing large
action IDs such as 3169590176.

This changes Action.ID, LinkAction.ID, and all actionID parameters
in service interfaces/implementations from int to int64.

Fixes digitalocean/terraform-provider-digitalocean#1532

BREAKING CHANGE: Action.ID, LinkAction.ID are now int64. All Get
methods accepting an actionID parameter now take int64 instead of int.

+++ storage_actions.go
@@ -12,7 +12,7 @@ import (
 type StorageActionsService interface {
 	Attach(ctx context.Context, volumeID string, dropletID int) (*Action, *Response, error)
 	DetachByDropletID(ctx context.Context, volumeID string, dropletID int) (*Action, *Response, error)
-	Get(ctx context.Context, volumeID string, actionID int) (*Action, *Response, error)
+	Get(ctx context.Context, volumeID string, actionID int64) (*Action, *Response, error)
 	List(ctx context.Context, volumeID string, opt *ListOptions) ([]Action, *Response, error)
 	Resize(ctx context.Context, volumeID string, sizeGigabytes int, regionSlug string) (*Action, *Response, error)
 }
@@ -48,7 +48,7 @@ func (s *StorageActionsServiceOp) DetachByDropletID(ctx context.Context, volumeI
 }
 
 // Get an action for a particular storage volume by id.
-func (s *StorageActionsServiceOp) Get(ctx context.Context, volumeID string, actionID int) (*Action, *Response, error) {
+func (s *StorageActionsServiceOp) Get(ctx context.Context, volumeID string, actionID int64) (*Action, *Response, error) {
 	path := fmt.Sprintf("%s/%d", storageAllocationActionPath(volumeID), actionID)
 	return s.get(ctx, path)
 }
